# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: MyPay fintech API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/users:
    get:
      operationId: adminListUsers
      tags: [admin]
      summary: List all users
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: before
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUsersResult"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get full user profile
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserDetail"
  /admin/users/{id}/suspend:
    patch:
      operationId: adminSuspendUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonInput"
      responses:
        "200":
          description: Suspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/unsuspend:
    patch:
      operationId: adminUnsuspendUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unsuspended
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/block:
    patch:
      operationId: adminBlockUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonInput"
      responses:
        "200":
          description: Blocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/unblock:
    patch:
      operationId: adminUnblockUser
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unblocked
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/wallet/freeze:
    patch:
      operationId: adminFreezeWallet
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonInput"
      responses:
        "200":
          description: Frozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/wallet/unfreeze:
    patch:
      operationId: adminUnfreezeWallet
      tags: [admin]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Unfrozen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

  # ── ADMIN DEPOSITS ───────────────────────────────────────────
...
components:
  schemas:
...
    SuccessResponse:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
        message:
          type: string

    # ── Auth schemas ──────────────────────────────────────────
...
    AdminUserSummary:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        mobile:
          type: string
        referralCode:
          type: string
        isActive:
          type: boolean
        isSuspended:
          type: boolean
        isBlocked:
          type: boolean
        inrBalance:
          type: ["string", "null"]
        walletFrozen:
          type: ["boolean", "null"]
        createdAt:
          type: string
    AdminUsersResult:
      type: object
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/AdminUserSummary"
        hasMore:
          type: boolean
        nextCursor:
          type: ["string", "null"]
    AdminUserDetail:
      type: object
      properties:
        user:
          type: object
    ReasonInput:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
...